pyproject.toml
[project]
name = "rogueedit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.5",
    "tcod>=19.3.0",
]
